fix: Fix request changes
- Separate the image processing function from your route file
- Add a prettier script
- Fix the endpoint in your readme file, so it can be a "provided endpoint that gives back status 200"
- Add typed parameters and return values for your route function
