+++ 01-image-processing-api/package.json
@@ -4,7 +4,8 @@
   "description": "Image Processing",
   "main": "index.js",
   "scripts": {
-    "lint": "eslint '*/**/*.{js,ts,tsx}' --quiet --fix",
+    "lint": "eslint '*/**/*.ts'",
+    "prettier": "prettier '*/**/*.ts' --write",
     "build": "npx tsc",
     "jasmine": "jasmine",
     "test": "npm run build && npm run jasmine",
