+++ 01-image-processing-api/src/routes/index.ts
@@ -3,10 +3,10 @@ import images from './api/images';
 
 const routes = express.Router();
 
-routes.get('/', (req, res) => {
+routes.get('/', (req: express.Request, res: express.Response): void => {
   res.send('main api routes');
 });
 
-routes.use('/images', images.images);
+routes.use('/images', images);
 
 export default routes;
